+++ index.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html data-bs-theme="light">
+    <head>
+        <title></title>
+        <link rel="stylesheet" href="style/style.css">
+        <meta name="viewport" content="width=device-width, initial-scale=1">
+    </head>
+    <body>
+        <div id="main"></div>
+        <script type="text/javascript" src="js/darkmode.js"></script>
+    </body>
+</html>
