+++ src/environment/mod.rs
@@ -1,9 +1,9 @@
 use dioxus::prelude::Component;
 
 #[cfg(not(target_family = "wasm"))]
-// mod desktop;
+mod desktop;
 
-// #[cfg(target_family = "wasm")]
+#[cfg(target_family = "wasm")]
 mod web;
 
 #[cfg(all(not(target_family = "wasm"), debug_assertions))]
@@ -11,9 +11,9 @@ mod hot_reload;
 
 pub fn init(root: Component) {
     #[cfg(not(target_family = "wasm"))]
-    // desktop::init_app(root);
+    desktop::init_app(root);
 
-    // #[cfg(target_family = "wasm")]
+    #[cfg(target_family = "wasm")]
     web::init_app(root);
 }
